server/pa-server/requirements.txt
asgiref==3.2.3
botocore==1.15.41
Django==3.0.2
django-cors-headers==3.2.1
django-storages==1.9.1
docutils==0.15.2
jmespath==0.9.5
mysqlclient==1.4.6
numpy==1.18.1
opencv-contrib-python-headless==4.2.0.34
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
